+++ packages/orq-rc/src/funcs/datasetsListDatapoints.ts
@@ -18,6 +18,7 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
+import * as errors from "../models/errors/index.js";
 import { OrqError } from "../models/errors/orqerror.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
@@ -38,6 +39,7 @@ export function datasetsListDatapoints(
 ): APIPromise<
   Result<
     operations.ListDatasetDatapointsResponseBody,
+    | errors.HonoApiError
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -63,6 +65,7 @@ async function $do(
   [
     Result<
       operations.ListDatasetDatapointsResponseBody,
+      | errors.HonoApiError
       | OrqError
       | ResponseValidationError
       | ConnectionError
@@ -152,8 +155,13 @@ async function $do(
   }
   const response = doResult.value;
 
+  const responseFields = {
+    HttpMeta: { Response: response, Request: req },
+  };
+
   const [result] = await M.match<
     operations.ListDatasetDatapointsResponseBody,
+    | errors.HonoApiError
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -164,9 +172,10 @@ async function $do(
     | SDKValidationError
   >(
     M.json(200, operations.ListDatasetDatapointsResponseBody$inboundSchema),
+    M.jsonErr(404, errors.HonoApiError$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
-  )(response, req);
+  )(response, req, { extraFields: responseFields });
   if (!result.ok) {
     return [result, { status: "complete", request: req, response }];
   }