+++ packages/orq-rc/RELEASES.md
@@ -10229,3 +10229,13 @@ Based on:
 - [typescript v4.11.0-rc.44] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.44] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.44 - packages/orq-rc
+
+## 2026-06-24 13:27:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.785.0 (2.912.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.45] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.45 - packages/orq-rc